Fix a Windows burn from the last commit. bugref:9376: Complete hardware cursor implementation in VMSVGA.
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@77133 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Main/src-client/DisplayImpl.cpp
@@ -4453,7 +4453,7 @@ DECLCALLBACK(void) Display::i_displayVBVAInputMappingUpdate(PPDMIDISPLAYCONNECTO
     pThis->i_handleUpdateVBVAInputMapping(xOrigin, yOrigin, cx, cy);
 }
 
-DECLCALLBACK(void) Display::i_displayVBVAReportCursorPosition(PPDMIDISPLAYCONNECTOR pInterface, uint32_t fFlags, ULONG aScreenId, uint32_t x, uint32_t y)
+DECLCALLBACK(void) Display::i_displayVBVAReportCursorPosition(PPDMIDISPLAYCONNECTOR pInterface, uint32_t fFlags, uint32_t aScreenId, uint32_t x, uint32_t y)
 {
     LogFlowFunc(("\n"));
